requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
Flask==3.0.2
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
pipreqs==0.4.13
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
yarg==0.1.9
